+++ lib/generators/react_on_rails/templates/redux/base/app/javascript/bundles/HelloWorld/components/HelloWorld.jsx
@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import style from './HelloWorld.module.css';
 
 const HelloWorld = ({ name, updateName }) => (
   <div>
@@ -9,7 +10,7 @@ const HelloWorld = ({ name, updateName }) => (
     </h3>
     <hr />
     <form>
-      <label htmlFor="name">
+      <label className={style.bright} htmlFor="name">
         Say hello to:
         <input id="name" type="text" value={name} onChange={(e) => updateName(e.target.value)} />
       </label>