+++ spec/dummy/config/webpack/test.js
@@ -1,8 +1,8 @@
-process.env.NODE_ENV = process.env.NODE_ENV || 'test';
+process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
 const webpackConfig = require('./webpackConfig');
 
-const testOnly = () => {
+const testOnly = (_clientWebpackConfig, _serverWebpackConfig) => {
   // place any code here that is for test only
 };